requirements.txt
PySide6==6.5.0
PySide6-Addons==6.5.0
PySide6-Essentials==6.5.0
python-slugify==8.0.1
pytube==15.0.0
shiboken6==6.5.0
text-unidecode==1.3
